--- src/main/java/com/example/demo/auth/controller/TokenApiController.java
@@ -1,25 +1,0 @@
-package com.example.demo.auth.controller;
-
-import com.example.demo.auth.dto.CreateAccessTokenRequest;
-import com.example.demo.auth.dto.CreateAccessTokenResponse;
-import com.example.demo.auth.service.TokenService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor(access = lombok.AccessLevel.PROTECTED)
-@RestController
-public class TokenApiController {
-    private final TokenService tokenService;
-
-    @PostMapping("/api/token")
-    public ResponseEntity<CreateAccessTokenResponse> createNewAccessToken
-            (@RequestBody CreateAccessTokenRequest request) {
-        String newAccessToken = tokenService.createNewAccessToken(request.getRefreshToken());
-
-        return ResponseEntity.status(HttpStatus.CREATED).body(new CreateAccessTokenResponse(newAccessToken));
-    }
-}
