--- src/main/java/com/example/demo/auth/dto/CreateAccessTokenRequest.java
@@ -1,10 +1,0 @@
-package com.example.demo.auth.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CreateAccessTokenRequest {
-    private String refreshToken;
-}
