--- src/main/java/com/example/demo/auth/dto/CreateAccessTokenResponse.java
@@ -1,10 +1,0 @@
-package com.example.demo.auth.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class CreateAccessTokenResponse {
-    private String accessToken;
-}
