--- src/main/java/com/example/demo/auth/repository/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package com.example.demo.auth.repository;
-
-import com.example.demo.auth.entity.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-    Optional<RefreshToken> findByUserId(Long userId);
-    Optional<RefreshToken> findByRefreshToken(String refreshToken);
-}
