--- src/main/java/com/example/demo/auth/service/TokenService.java
@@ -1,29 +1,0 @@
-package com.example.demo.auth.service;
-
-import com.example.demo.config.jwt.TokenProvider;
-import com.example.demo.user.entity.User;
-import com.example.demo.user.service.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.time.Duration;
-
-@RequiredArgsConstructor(access = lombok.AccessLevel.PROTECTED)
-@Service
-public class TokenService {
-
-    private final TokenProvider tokenProvider;
-    private final RefreshTokenService refreshTokenService;
-    private final UserService userService;
-
-    public String createNewAccessToken(String refreshToken) {
-        if(!tokenProvider.validToken(refreshToken)) {
-            throw new IllegalArgumentException("유효하지 않은 리프레시 토큰입니다.");
-        }
-
-        Long userId = refreshTokenService.findByRefreshToken(refreshToken).getUserId();
-        User user = userService.findById(userId);
-
-        return tokenProvider.generateToken(user, Duration.ofHours(2));
-    }
-}
