--- src/main/java/com/example/demo/config/jwt/JwtProperties.java
@@ -1,19 +1,0 @@
-package com.example.demo.config.jwt;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-/**
- * JwtProperties
- * application.yml 에서 jwt 설정 정보를 가져와서 담는 클래스
- * */
-@Setter
-@Getter
-@Component
-@ConfigurationProperties("jwt")
-public class JwtProperties {
-    private String issuer;
-    private String secretKey;
-}
