+++ src/main/java/com/example/demo/domain/auth/dto/CreateAccessTokenRequest.java
@@ -1,0 +1,10 @@
+package com.example.demo.domain.auth.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CreateAccessTokenRequest {
+    private String refreshToken;
+}
