+++ src/main/java/com/example/demo/domain/auth/dto/CreateAccessTokenResponse.java
@@ -1,0 +1,10 @@
+package com.example.demo.domain.auth.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class CreateAccessTokenResponse {
+    private String accessToken;
+}
