+++ src/main/java/com/example/demo/domain/auth/repository/RefreshTokenRepository.java
@@ -1,0 +1,11 @@
+package com.example.demo.domain.auth.repository;
+
+import com.example.demo.domain.auth.entity.RefreshToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
+    Optional<RefreshToken> findByUserId(Long userId);
+    Optional<RefreshToken> findByRefreshToken(String refreshToken);
+}
