+++ src/main/java/com/example/demo/domain/user/controller/UserController.java
@@ -1,0 +1,15 @@
+package com.example.demo.domain.user.controller;
+
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class UserController {
+    @GetMapping("/api/me")
+    public String me() {
+        return "me";
+    }
+}
