+++ src/main/java/com/example/demo/domain/user/controller/UserViewController.java
@@ -1,0 +1,12 @@
+package com.example.demo.domain.user.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserViewController {
+    @GetMapping("/login")
+    public String login() {
+        return "oauthLogin";
+    }
+}
