--- src/main/java/com/example/demo/user/controller/UserController.java
@@ -1,13 +1,0 @@
-package com.example.demo.user.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class UserController {
-
-    @GetMapping("/api/me")
-    public String me() {
-        return "me";
-    }
-}
