--- src/main/java/com/example/demo/user/controller/UserViewController.java
@@ -1,12 +1,0 @@
-package com.example.demo.user.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class UserViewController {
-    @GetMapping("/login")
-    public String login() {
-        return "oauthLogin";
-    }
-}
